requirements.txt
aiogram
python-dotenv
loguru